requirements.txt
fastapi[standard]
line-bot-sdk
openai
PyYAML